refactor: move lock period display from RewardsCard to StakingAdditionalInfo component
