+++ src/components/DaoStaking/StakingAdditionalInfo.tsx
@@ -1,9 +1,18 @@
 import { Flex, Text } from '@chakra-ui/react';
 import { Info } from '@phosphor-icons/react';
 import { useTranslation } from 'react-i18next';
+import { useDurationDisplay } from '../../helpers/dateTime';
+import { useCurrentDAOKey } from '../../hooks/DAO/useCurrentDAOKey';
+import { useDAOStore } from '../../providers/App/AppProvider';
 
 export function StakingAdditionalInfo() {
   const { t } = useTranslation('staking');
+  const { daoKey } = useCurrentDAOKey();
+  const {
+    governance: { stakedToken },
+  } = useDAOStore({ daoKey });
+
+  const lockPeriod = useDurationDisplay(stakedToken?.minimumStakingPeriod);
 
   return (
     <Flex
@@ -12,6 +21,31 @@ export function StakingAdditionalInfo() {
       gap={2}
       alignSelf="stretch"
     >
+      <Flex
+        direction="column"
+        alignItems="flex-start"
+        gap="8px"
+        alignSelf="stretch"
+      >
+        <Flex
+          justifyContent="space-between"
+          alignItems="flex-start"
+          alignSelf="stretch"
+        >
+          <Text
+            color="color-content-content1-foreground"
+            textStyle="text-sm-regular"
+          >
+            {t('lockPeriod')}
+          </Text>
+          <Text
+            color="color-content-content4-foreground"
+            textStyle="text-sm-regular"
+          >
+            {lockPeriod}
+          </Text>
+        </Flex>
+      </Flex>
       <Flex
         padding="12px"
         alignItems="flex-start"
